Timing für Animationen eingefügt FPS Counter eingebaut

+++ src/app/ledprojekt/Entity.java
@@ -22,8 +22,8 @@ public interface Entity extends Collidable {
     void updateKeyEventRef(KeyEvent event);
     KeyEvent getKeyEventRef();
 
-    void updateDelta(float dt);
-    float getDelta();
+    void updateDelta(double dt);
+    double getDelta();
 
     void draw(BoardController controller);
 }
